+++ client/my-sites/email/gsuite-purchase-cta/test/index.js
@@ -13,10 +13,7 @@ import { createReduxStore } from 'state';
 import GSuitePurchaseCta from '../';
 
 jest.mock( 'components/email-verification/email-verification-gate', () => 'EmailVerificationGate' );
-jest.mock(
-	'my-sites/domains/domain-management/gsuite/gsuite-purchase-features',
-	() => 'GSuitePurchaseFeatures'
-);
+jest.mock( 'my-sites/email/gsuite-purchase-features', () => 'GSuitePurchaseFeatures' );
 
 describe( 'GSuitePurchaseCta', () => {
 	test( 'it renders GSuitePurchaseCta with basic plan', () => {
