+++ client/my-sites/email/gsuite-purchase-features/style.scss
